backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Quantitative Finance
vectorbt==0.26.0
ta-lib==0.4.28
backtrader==1.9.78.123

# Google Gemini AI
google-generativeai==0.3.2

# Supabase (Vector DB)
supabase==2.3.0
vecs==0.4.0

# Exchange APIs
python-binance==1.0.19
alpaca-trade-api==3.0.2
polygon-api-client==1.12.5

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Telegram Bot
python-telegram-bot==20.7

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
